requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langchain-huggingface
beautifulsoup4
python-dotenv
faiss-cpu
sentence-transformers
redis
langchain-redis